refactor: move transaction script to single component
